pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rollcall"
version = "0.1.0"
description = "RollCall: OCR end credits with Gemini and rename unlabeled media files"
readme = "README.md"
requires-python = ">=3.9"                
dependencies = [
  "ffmpeg-python>=0.2.0",
  "Pillow>=10.0.0",
  "google-genai>=1.0.0",                 
  "typer>=0.12.3",
]

[project.scripts]
rlcl = "rollcall.cli:entrypoint"

[tool.setuptools]
packages = ["rollcall"]
